+++ src/lib/io/live/marketplace.ts
@@ -1,10 +1,10 @@
 import { DateTime, Duration, Interval } from 'luxon';
 import fetch from 'node-fetch';
-import cache from '../../io/cache.js';
-import { MarketplaceService, Progress } from '../../io/interfaces.js';
 import { RawLicense, RawTransaction } from "../../model/marketplace/raw";
 import env from '../../parameters/env.js';
 import { AttachableError } from '../../util/errors.js';
+import cache from '../cache.js';
+import { MarketplaceService, Progress } from '../interfaces.js';
 
 export class LiveMarketplaceService implements MarketplaceService {
